backend/requirements.txt
# Core API Framework
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Pydantic & Settings
pydantic==2.10.3
pydantic-settings==2.7.0

# Supabase
supabase==2.11.0
storage3>=0.8.2

# Data Processing
pandas==2.2.3
numpy>=1.26.0
scipy>=1.11.0

# Machine Learning
tensorflow==2.18.0
scikit-learn==1.6.1
keras==3.8.0
catboost==1.2.7
xgboost==2.1.3
lightgbm==4.5.0
imbalanced-learn>=0.11.0

# HTTP Client
httpx==0.28.1

# Environment Variables
python-dotenv==1.0.1

# CORS
starlette>=0.37.2

# Production Server (for deployment)
gunicorn==23.0.0

# Serverless Adapter (for Vercel)
mangum==0.17.0

# Utilities
python-dateutil==2.9.0.post0
